+++ packages/terra-list/src/terra-dev-site/test/list/ListSection.test.jsx
@@ -2,12 +2,13 @@ import React from 'react';
 import List, { Item, Section } from 'terra-list/lib/index';
 
 const ListSectionTest = () => (
-  <List>
+  <List role="listbox" aria-label="list section">
     <Section
       key="static-section"
       title="Static Section"
+      role="option"
     >
-      <Item key="123">
+      <Item key="123" isSelectable>
         <p>test static</p>
       </Item>
     </Section>
@@ -16,7 +17,7 @@ const ListSectionTest = () => (
       isCollapsible
       title="Collapsible Section"
     >
-      <Item key="223">
+      <Item key="223" isSelectable isDisabled>
         <p>test collapsible</p>
       </Item>
     </Section>