+++ packages/terra-list/tests/jest/ListSection.test.jsx
@@ -3,13 +3,13 @@ import { Item, Section } from '../../src/index';
 
 // Snapshot Tests
 it('should render with items', () => {
-  const item1 = <Item key="123" />;
-  const item2 = <Item key="124" />;
-  const item3 = <Item key="125" />;
-  const item4 = <Item key="126" />;
-  const item5 = <Item key="127" />;
+  const item1 = <Item isSelectable key="123" />;
+  const item2 = <Item isSelectable key="124" />;
+  const item3 = <Item isSelectable isDisabled key="125" />;
+  const item4 = <Item isSelectable isDisabled key="126" />;
+  const item5 = <Item isSelectable key="127" />;
   const items = [item1, item2, item3, item4, item5];
-  const shallowComponent = shallow(<Section title="test">{items}</Section>);
+  const shallowComponent = shallow(<Section role="listbox" title="test">{items}</Section>);
   expect(shallowComponent).toMatchSnapshot();
 });