+++ packages/terra-action-header/src/ActionHeader.jsx
@@ -14,7 +14,8 @@ const propTypes = {
    */
   children: PropTypes.element,
   /**
-   * Optionally sets the heading level. One of `1`, `2`, `3`, `4`, `5`, `6`. Default `level=1`.
+   * Optionally sets the heading level. One of `1`, `2`, `3`, `4`, `5`, `6`. Default `level=1`. This helps screen readers to announce appropriate heading levels.
+   * Changing 'level' will not visually change the style of the content.
    */
   level: PropTypes.oneOf([1, 2, 3, 4, 5, 6]),
   /**