+++ packages/terra-list/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Changed
 * `isDisabled` prop added to fix accessibility issue to handle selectable and non-selectable items correctly.
+* `role` prop added to list sections, subsection, sectionHeader and subsectionHeader to fix accessibility issues.
 
 4.25.0 - (January 28, 2020)
 ------------------