+++ packages/terra-list/CHANGELOG.md
@@ -3,6 +3,13 @@ ChangeLog
 
 Unreleased
 ----------
+### Added
+* Added clinical-lowlight-theme theming files
+
+4.25.0 - (January 28, 2020)
+------------------
+### Changed
+* Minor dependency version bump
 
 4.24.0 - (January 7, 2020)
 ------------------